% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dysprop.R
\name{cos_disprop}
\alias{cos_disprop}
\title{Wspolczynnik 1-cos.}
\usage{
cos_disprop(x, y, pierwiastek = FALSE, komponent = NULL)
}
\arguments{
\item{x}{wektory liczbowy}

\item{y}{wektory liczbowy}

\item{pierwiastek}{wartość logiczna - czy wyciągnąć pierwiastek ze zwracanej wartości
współczynnika (uzyskując w efekcie zmodyfikowany współczynnik Gallaghera)?}

\item{komponent}{opcjonalnie wektor definiujący podział na komponenty}
}
\value{
Funkcja zwraca wektor (jednoelementowy gdy nie jest przeprowadzana dekompozycja
lub (liczba_grup + 1)-elementowy, gdy jest przeprowadzana) z wartościami współczynnika (i ew. jego
komponentów).
}
\description{
Wylicza współczynnik 1-cos, zaproponowany przez Koppela i Diskina.
Zaimplementowano także możliwość przeprowadzenia dekompozycji na komponenty
w postaci analogicznej do zaproponowanej przez Powella and Tuckera.
}
\details{
Uwaga, porównywane wektory nie mogą zawierać braków danych!

Dekompozycja może być przeprowadzona tylko w sytuacji, gdy parametr \code{pierwiastek}
przyjmuje wartość \code{FALSE}.
}
\examples{
vTG=c(10,40,50)
sTG=c(15,25,60)
cos_disprop(vTG, sTG)
cos_disprop(vTG, sTG, komponent=c(1, 2, 2))
}

