% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dysprop.R
\name{dissim}
\alias{dissim}
\title{Wspolczynnik D (Duncana, Pedersena, or Loosemore-Hanby'ego).}
\usage{
dissim(x, y, komponent = NULL)
}
\arguments{
\item{x}{wektory liczbowy}

\item{y}{wektory liczbowy}

\item{komponent}{opcjonalnie wektor definiujący podział na komponenty}
}
\value{
Funkcja zwraca wektor (jednoelementowy gdy nie jest przeprowadzana dekompozycja
lub (liczba_grup + 1)-elementowy, gdy jest przeprowadzana) z wartościami współczynnika (i ew. jego
komponentów).
}
\description{
Wylicza współczynnik D (Duncana, Pedersena, or Loosemore-Hanby'ego).
Zaimplementowano także możliwość przeprowadzenia dekompozycji na dwa komponenty
w postaci zaproponowanej przez Powella and Tuckera.
}
\details{
Uwaga, porównywane wektory nie mogą zawierać braków danych!
}
\examples{
vTG=c(10,40,50)
sTG=c(15,25,60)
dissim(vTG, sTG)
dissim(vTG, sTG, c(1, 2, 2))
}

