% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dysprop.R
\name{divergence_kl}
\alias{divergence_kl}
\title{Dywergencja Kullbacka-Leiblera (entropia wzgledna).}
\usage{
divergence_kl(x, y, p = 2)
}
\arguments{
\item{x}{wektor ważony}

\item{y}{wektor ważący}

\item{p}{podstawa logarytmu}
}
\value{
Funkcja zwraca wartość liczbową.
}
\description{
Wylicza dywergencje Kullbacka-Leiblera (entropię względną).
}
\details{
Uwaga, porównywane wektory nie mogą zawierać braków danych!
}
\examples{
vTG=c(10,40,50)
sTG=c(15,25,60)
divergence_kl(sTG, vTG)
}

