% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dysprop.R
\name{gallagher}
\alias{gallagher}
\title{Wspolczynnik Gallaghera.}
\usage{
gallagher(x, y, doKwadratu = FALSE, komponent = NULL)
}
\arguments{
\item{x}{wektory liczbowy}

\item{y}{wektory liczbowy}

\item{doKwadratu}{wartość logiczna - czy podnieść zwracaną wartość współczynnika
Gallaghera do kwadratu?}

\item{komponent}{opcjonalnie wektor definiujący podział na komponenty}
}
\value{
Funkcja zwraca wektor (jednoelementowy gdy nie jest przeprowadzana dekompozycja
lub (liczba_grup + 1)-elementowy, gdy jest przeprowadzana) z wartościami współczynnika (i ew. jego
komponentów).
}
\description{
Wylicza współczynnik Gallaghera.
Zaimplementowano także możliwość przeprowadzenia dekompozycji na komponenty
w postaci analogicznej do zaproponowanej przez Powella and Tuckera.
}
\details{
Uwaga, porównywane wektory nie mogą zawierać braków danych!

Dekompozycja może być przeprowadzona tylko w sytuacji, gdy parametr \code{doKwadratu}
przyjmuje wartość \code{TRUE}.
}
\examples{
vTG=c(10,40,50)
sTG=c(15,25,60)
gallagher(vTG, sTG)
gallagher(vTG, sTG, komponent=c(1, 2, 2))
}

