% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/podzial.R
\name{podziel}
\alias{podziel}
\title{Podzial mandatow (i ew. innych rzeczy)}
\usage{
podziel(x, m, regula)
}
\arguments{
\item{x}{wektor z liczbą głosów (wag)}

\item{m}{liczba mandatów do podziału}

\item{regula}{ciąg znaków definiujacy metodę podziału; jeden z: "D'Hondt", "Hamilton", "Hare-Niemeyer", "Jefferson", "Saint-League", "Webster", "DH", "H", "HN", "J", "SL", "W"}
}
\value{
funkcja zwraca wektor liczb całkowitych o długości równej długości wektora \code{x}
}
\description{
Funkcja wylicza podział mandatów (ew. innych niepodzielnych rzeczy) na podstawie rozkładu głosów (ew. innych wag).
}
\details{
Równoważne sobie wartości parametru \code{regula}:
\itemize{
\item "Saint-League", "Webster", "SL", "W";
\item "D'Hondt", "Jefferson", "DH", "J";
\item "Hamilton", "Hare-Niemeyer", "H", "HN".
}
}
\examples{
v1 = c(47, 16, 15.8, 12, 6.1, 3.1) * 1000
podziel(v1, 10, "SL")
podziel(v1, 10, "DH")
podziel(v1, 10, "HN")
# paradoks Alabamy
vA = c(15, 15, 9, 5, 5, 2) * 100
podziel(vA, 25, "HN")
podziel(vA, 26, "HN")
}

